Feature/f 1/rework part 2

+++ src/modules/albums/albums.controller.ts
@@ -5,46 +5,86 @@ import {
   Get,
   HttpCode,
   HttpStatus,
+  NotFoundException,
   Param,
   ParseUUIDPipe,
   Post,
   Put,
 } from '@nestjs/common';
 import { AlbumsService } from './albums.service';
-import { CreateAlbumDto } from './dto/create-album.dto';
-import { UpdateAlbumDto } from './dto/update-album.dto';
+import { ArtistsService } from '../artists/artists.service';
+import { CreateAlbumDto, UpdateAlbumDto } from './dto';
+import { Album } from 'prisma/prisma-client';
 
 @Controller('album')
 export class AlbumsController {
-  constructor(private readonly albumsService: AlbumsService) {}
+  constructor(
+    private readonly albumsService: AlbumsService,
+    private readonly artistsService: ArtistsService,
+  ) {}
 
   @Post()
   @HttpCode(HttpStatus.CREATED)
-  async create(@Body() createAlbumDto: CreateAlbumDto) {
+  async create(@Body() createAlbumDto: CreateAlbumDto): Promise<Album> {
+    if (createAlbumDto.artistId) {
+      const artist = await this.artistsService.findOne(createAlbumDto.artistId);
+      if (!artist) {
+        throw new NotFoundException('Artist with specified id is not found');
+      }
+    }
     return await this.albumsService.create(createAlbumDto);
   }
 
   @Get()
-  async findAll() {
+  async findAll(): Promise<Album[]> {
     return await this.albumsService.findAll();
   }
 
   @Get(':id')
-  async findOne(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    return await this.albumsService.findOne(id);
+  async findOne(
+    @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
+  ): Promise<Album> {
+    const album = await this.albumsService.findOne(id);
+
+    if (!album) {
+      throw new NotFoundException('Album with specified id not found');
+    }
+
+    return album;
   }
 
   @Put(':id')
   async update(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
     @Body() updateAlbumDto: UpdateAlbumDto,
-  ) {
+  ): Promise<Album> {
+    if (updateAlbumDto.artistId) {
+      const artist = await this.artistsService.findOne(updateAlbumDto.artistId);
+      if (!artist) {
+        throw new NotFoundException('Artist with specified id is not found');
+      }
+    }
+
+    const updatedAlbum = await this.albumsService.findOne(id);
+
+    if (!updatedAlbum) {
+      throw new NotFoundException('Album with specified id not found');
+    }
+
     return await this.albumsService.update(id, updateAlbumDto);
   }
 
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  async remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
+  async remove(
+    @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
+  ): Promise<Album> {
+    const removedAlbum = await this.albumsService.findOne(id);
+
+    if (!removedAlbum) {
+      throw new NotFoundException('Album with specified id not found');
+    }
+
     return await this.albumsService.remove(id);
   }
 }
